Update tests to not log in in before(:all) blocks
Also changes the API client to log out after creating an access token
if it wasn't already logged in. This is so that it doesn't leave the
browser logged in at the start of tests when they expect to be able to
log in. For example, if a test is the first in the suite it will log
in to create an access token. If that happens in a before(:context)
block, the test will already be logged in, so an attempt to log in
would fail.