Allow AWS role auth to update an existing role
Now that AWS credentials are passed to the frontend
for the EKS form, we need to be able to re-authorize
an existing role to generate a new set of credentials
(in the event that the previous set expired).

This will be a no-op most of the time (as the user
simply needs credentials for their role without changing
it), but the role will be updated if requested.