Remove validate_merge_sha feature flag
This removes the `validate_merge_sha` feature flag that was added in
https://gitlab.com/gitlab-org/gitlab/merge_requests/19579 which was
included in 12.5

Starting 12.5 we started writing the merge_sha inside the merge params
for validating.

To avoid merges failing during a deploy, the feature flag wasn't
defaulted to "on". But one release later we can remove it since the
SHAs will would have been written in the params regardless.