Abort rendering of disabled security reports
This MR creates a new EE::MergeRequest#enabled_reports method that
indicates what MR reports are enabled/disabled. We then use this data
to populate the MergeRequestStore in order to abort the rendering of
disabled reports in the MR widget instead of allowing the reports to
show up and eventually error out when the corresponding XHR requests
resolve.

This change only affects reports rendering when the corresponding
*MergeRequestReportApi feature flag is enabled.