Code review cleanup & spec fixes
Upated store_mentions! method to return true if no exception
is raised on mentions saving. That is needed in order to be
able  to wrap it in a transaction with its respective model.save,
if both model.save and model.store_mentions! are successfull, then
model is considered to be saved.