Add option to delete cached Kubernetes namespaces
If a user modifies their Kubernetes cluster directly
it can become out of sync with what GitLab has stored,
which can cause deploys to fail.

By deleting these records we allow them to be fetched
from the cluster on the next deploy, which brings
GitLab back in sync with the cluster.

https://gitlab.com/gitlab-org/gitlab/merge_requests/20411