Refactor snippet route helpers
In several places of the code we have conditions to check
whether the snippet is a `PersonalSnippet` or a `ProjectSnippet`.
In this commit we overwrite existing named routes helpers to
avoid that redundancy and in preparation for secret snippets,
where we will have to automatically add a query string param
to every personal snippet route.