Fix project creation with templates using /projects/user/:id API
Previously creation with an instance or group template with POST
`/api/v4/projects worked`, but `/api/v4/projects/user/:id` did not
because the latter did not accept the `use_custom_templates` parameter
due to an oversight.

This commit fixes this by adding the `optional_create_project_params`
Grape API and add tests for project creation via the user endpoint.

Closes https://gitlab.com/gitlab-org/gitlab/issues/37004