Assume can sign in if sign in button is present
The standard sign in process has been replaced with an experimental
one that excludes the Sign In tab when there's only the one standard
sign in method available.

Rather than check for the Sign In tab, this change assumes it's
possible to sign in again if the sign button is present. We don't
really care what sign in options are available, as long as we were
signed out successfully.