Fix error handling by using a backoff timeout and a flash notice
When a metric cannot be loaded, our frontend would stop after 3 tries
if a metric loading failed. This change removes the retry count and
replaces it with a long timeout.

This is done by grouping all dashboard metrics requests in a
Promise.all and by prompting the user when any of them fails.

The other metrics can still load, so specs are added to emulate
these condition when a single metric fails have been updated.

Add changelog file, remove unused gitlab.pot string