Re-enable RequestStore by default in Sidekiq
https://gitlab.com/gitlab-org/gitlab/merge_requests/20291 accidentally
made the `RequestStore` Sidekiq middleware disabled by default instead
of enabled by default. This will cause a performance regression in
Sidekiq jobs. For example, the Markdown cache will not be used.

For now, we default the `REQUEST_STORE` environment variable to 1. In
the future, we should consider just dropping this environment variable.

Discovered in
https://gitlab.com/gitlab-org/gitlab/merge_requests/19759#note_250602320