Send results only for external_pipelines
On GitHub integrations we might have multiple pipelines for the
same SHA. One pipeline triggered by the commit and a different
one from the pull request event. On GitHub, we update the status
of the pull request with the status of the pipeline that finishes
last. This commit cancel sending messages if we also have a
external_pull_request_event pipeline for the same sha.