Validate deployment SHAs and refs
This prevents users from creating deployments using the API while
providing an invalid SHA or ref name. Providing invalid data for these
fields will break certain parts of GitLab, such as the UI used for
displaying the deployments of an environment.

These changes require that we correct a variety of tests that were using
deployments without a valid repository, or were even testing behaviour
that would never run due to the lack of a repository.

This fixes https://gitlab.com/gitlab-org/gitlab/issues/36967