Fix error in updating runner session
If a runner attempted to pass along `session` parameters, an error in
`Ci::RegisterJobService` would be encountered if an existing session
were present. This occurred because the `id` field did not exist in the
parameters, so Rails would signal an error:

```
Failed to remove the existing associated runner_session. The record
failed to save after its foreign key was set to nil.
```

By adding the `update_only: true`, we allow Rails to update an existing
session.

Closes https://gitlab.com/gitlab-org/gitlab/issues/34366