Excludes project-only services from the admin service templates area
Introduces project_only method to the Service class that defaults to
false. When overridden in a subclass that service will not appear in
the list of service templates that can be set instance-wide.

Fixes a bug where the Generic Alerts Integration appeared at the top
of the list and produced an exception when clicked.