Support diff version arguments for diffs batch endpoint
It's a next step towards deprecating the regular
`/diffs` endpoint in favor of `/diffs_batch`.

Now we can receive pagination arguments, even if we
fetch from the repository. Though, we won't
batch these results yet, which is a work for
https://gitlab.com/gitlab-org/gitlab/issues/32859.

This mainly avoid using two different endpoints for
persisted/batched diffs and repository/non-batch diffs.