Support different hostname for smartcard auth
This adds support for using an alternate host name for smartcard
authentication. Now the NGINX that handles TLS client authentication
can run on the same host (or a different host depending on the
deployment model) and use a different host name. In this case this
the server must be configured within GitLab NGINX. The limitation
is that the alternate host name must be a sub-domain of GitLab host
name.