Remove feature flag for limiting diverging commit counts
In GitLab 12.4 via
https://gitlab.com/gitlab-org/gitlab/merge_requests/16737, we required
the frontend to request at most 20 branches in the
`Projects::BranchesController#diverging_commit_counts` endpoint to avoid
Gitaly N+1 issues. This commit removes the feature flag and makes this
behavior mandatory. In the past week, there do not appear to be any
rejections, which would show up as 422 errors.