Update a test to avoid flakiness
This test has shown failures when running on CI, because automatic
comments sometimes do not show up, but with the two other expecta-
tions, we can still test that the functionality of assigning mul-
tiple users to an issue works.

So, better having one less expectation, but having reliable test
results, than having an extra expectation that leads to a flaky
test and false negatives.