Migrate from #multi to #pipelined for performance
Under the hood, #multi uses #pipelined, and we don't care about the
minor differences that using the #multi transaction approach gets us, so
we pick up about 0.1 sec of performance on large-ish diff collections.
This is a minor improvement, but as this is more or less a straight
drop-in replacement, and we're already workign directly with redis-rb,
it seems worthwhile.