Skip admin authentication for user oauth apps
Upgrading doorkeeper to 5.0.2 introduced a change where the method
`admin_authenticator` is required to be defined in the config, otherwise
all responses from the Doorkeeper will be a 403.

https://github.com/doorkeeper-gem/doorkeeper/wiki/Migration-from-old-versions#api-changes-3

How the admin authenticator behaves for our applications depends on the
context that the application is being created in - so a base level
definition of this will be complicated.

For now, skip this new before action in the controller. We already have
user level authentication in this context.