Signal-boost side effect for prepare method
The method does not only prepare the relation (as the `prepare_relation`
suggests), but also triggers batch counting which in turn executes the
query as a side-effect.

This is being signal-boosted with renaming the method to `prepare!`.

We use it in places where we actually want both (preloading/batch
counting) to happen. In other places (projects API), we actually don't
want that and have the ability to split this into multiple steps now.