Update security-harness for Security remote
Updates the pre-push hook to check for dev.gitlab.org *or* the
`gitlab-org/security/` remote.

In order to facilitate a smooth transition for users who already
installed the old version of the hook at some point, we check the
existing hook's SHA against a list of known previous versions.

If the user has a previous version of the payload installed, we simply
upgrade it in-place. Otherwise we fall back to the previous behavior of
warning the user that they have a hook we don't know about.

Further, this update adds colored terminal output in order to draw
attention to the various messages, but will respect a `NO_COLOR`
environment variable to disable this behavior.