Do not attribute unverified commit e-mails to GitLab users
Previously we would identify any user with a matching e-mail address as
a commit author or committer, but this caused confusion when commits
contained invalid, unconfirmable e-mail addresses (e.g. my.local).

Now, we only assign a commit author or committer is a specific user if
the e-mail has been confirmed.

Closes https://gitlab.com/gitlab-org/gitlab/issues/28493