Search for conflict headers everywhere
We seem to have restricted to only no haml, nor js, nor rb because we
can rely on tests. But since the new scripts is orders of magnitude
faster, we can include all files by default.

We need to exclude two files for now as these have conflict headers as
part of their tests.