Block specifying a k8s namespace via CI template for managed clusters
A namespace can only be specified via gitlab-ci.yml
for unmanaged clusters, as we currently have no way
of preventing a job requesting a namespace it
shouldn't have access to.

To make this clear, we fail the build instead of
silently using a namespace other than the one
explicitly specified.

https://gitlab.com/gitlab-org/gitlab/issues/27630