Validate connection section in direct upload config
For some reason, `Settingslogic` doesn't always provide the right error
message when the `connection` block is omitted. Sometimes it erroneously
attributes the error section to a different part of the YAML
(e.g. `backup`, `dependency_proxy`).

This commit adds on more information on which `connection` block is
missing from the `gitlab.yml` (e.g. LFS, artifacts, or uploads).

Closes https://gitlab.com/gitlab-org/gitlab/issues/38258