Fix pipeline retry in a CI DAG
Previously a retried pipeline for builds in a DAG would cause the newly-
created builds to sit in the created state because it would be waiting
for the dependencies to trigger it. However, in the retry case, the
dependencies may have already completed, so they will never trigger new
builds.

We fix this by:

1. Looking up successful builds with the required names
2. Passing those build IDs to `Ci::ProcessPipelineService` to trigger
   builds with those dependencies.

Closes https://gitlab.com/gitlab-org/gitlab/issues/36068