Spec forking a project after source moved shard
This specs forking a project that was already the source of a pool
repository after the source project was moved to a different shard.

In this case the old pool repository will continue to exist, containing
all the deduplicated objects from the forks on the first shard. But
we'll create a new PoolRepository for the shart the source project was
moved to.

Then, new forks will be deduplicated using the new PoolRepository on
the new shard.