Allow custom apply suggestion commit messages
This commit adds the capability to define a custom
commit message, which is used when a change
suggestion on a merge request is applied. The
configuration is done on a project settings level and
supports free text with GitLab variables. The old
hardcoded commit message is used as a fallback
option.