Fix incorrect diff file property reference
Before, the diffs *always* had both parallel and inline
diffs, so this should have never hit the else block.
Now that the split diffs feature is enabled, there are
cases (about 50% of the time) where the else is entered.

The test for inline diff lines was then trying to iterate over
the parallel lines, which of course don't exist in a lot of
situations.