Set GIT_DEPTH to 20
Before if the SHA of the last CI run were ahead of the next CI run by
more than 50 commits, the build would fail because Git assumes the full
history is present
(https://gitlab.com/gitlab-org/gitlab/issues/33041). It does not know
the current repository has a shallow clone.

With https://gitlab.com/gitlab-org/gitlab/issues/39134, the pre-clone
script now seeds the repository with the full history of `master` and a
consistent snapshot. As a result, we should be able to use a shallow
clone with a lower depth, which should decrease the load from the
server.