Integrate with single diff type feature flag
This commit pushes the single_mr_diff_view 
feature flag into the glFeatures JavaScript 
object as soon as the merge requests controller 
starts up.

Because that feature flag is needed in the Vuex 
actions and mutations, as soon as the Diffs Vue 
app starts, it also takes that feature flag and puts 
it into state.