Simplifies flaky MR review QA test
The test duplicated coverage provided by unit/feature tests, so this
change makes the test much simpler, covering only the happy path of
the MR review feature, plus additional coverage for an S1 regression.

Adds waits after a few actions so that the test doesn't continue to
another action while the previous is still not complete.

Sets `minimum` when finding elements by index because we know at least
how many elements to expect. And then we don't need an extra wait
before using `all` because `minimum` takes care of that - it will no
longer return an empty array if no elements are found, it will raise
an exception instead.