Create dedicated log file for Elasticsearch integration
Dedicated log files for different parts of the system will help us and
our users debugging parts of the system that obviously relate to this
feature.

In future we should add even more logging to this file especially around
indexing events and any exceptions we're catching in sidekiq or web.