Delete flaky end-to-end test
This commit deletes this test since it is flaky when executed on CI
, but I was never able to reproduce the issue locally.

Also, based on the following discussion, such feature is already
covered with lower-level tests, which makes it safe to delete it.

https://gitlab.com/gitlab-org/gitlab/issues/39595#note_258930120

The affected view and page object were updated accordingly, to re-
move code that was only used by the deleted test, so we don't leave
any dead code.