Extract list_projects into a separate controller
Extracts `ErrorTrackingController#list_projects` to
`ErrorTracking::ProjectsController`

The main goal of the change is moving towards RESTful controllers,
and have a separate controller per resource.
As well as unload the `ErrorTrackingController` by extracting more
code out of it.

It also follows an established practice to use the HTTP GET method
for reading data.

* Extract Sentry project resource into a separate controller
* Change HTTP method from POST to GET
* Update front-end part of using a new route