Add rolling median line to CA duration chart
In order to do this, we need to fetch data offset from
the selected dates. The offset will be equal to the number
of days that the data is being displayed for.

I.e. if we display data for 30 days, we will need to fetch
data for 60 days.

We are also introducing a feature flag specifically for
the median line.