Add warnings about changes: to CI Config docs
Customers have frequently attempted to use changes: policies in
scheduled and manually run pipelines that have no Git push event.
There are no new or old commits in this scenario, and changes:
returns true by default and causes a great deal of confusion. We
should be more vocal about what `changes:` means and why it works
the way it does.