Wait for elasticsearch server before searching
Checks for a valid search response via the API using an
elasticsearch-only scope. Checks 3 times and raises an exception if
unable to search.

This gives the elasticsearch server time to be ready for the the
index to be created before the first test.

Waiting for a response takes time, so it should be enough to make 3
attempts with the default interval between attempts.