Replace Node `joinPaths` with custom impl
**Why?**
The current implementation of `joinPaths` comes from the
Node core modules `path` and has a number of caveats:

- Does not support non-string values
- Does not handle joining absolute URL's

https://nodejs.org/api/path.html#path_path_join_paths

**Note:**
The new impl of `joinPaths` does not do the same special
normalization which the Node version did. This should be
safe because I'm pretty sure we only use this method to
join paths with `/` that potentially start and end with
a `/`.