Make `workflow:rules` to work well with Merge Requests
This improves `workflow:rules` behavior to allow
workflow rules to control when pipeline is created.

For, example it allows to define a global workflow
to create pipelines on a branch, and on a merge request
in an easy way without the need to re-invent that multiple
times for each job.