Skip LFS updates in mirrors if repository has not changed
The GetAllLFSPointers RPC is expensive and was being called for every
mirror. If the repository has not changed, we can assume that the LFS
pointers also did not change. We can skip calling this RPC if there are
no updates.

Since we don't actually get a return code if branches or tags were
updated, we use the repository checksum as a cheaper mechanism (which
XORs the ref contents) to determine whether we run the LFS scan.

Closes https://gitlab.com/gitlab-org/gitlab/issues/38256