Always check override where prepend happens
Because otherwise helpers won't be verified, and since we're using
prepend to override direct methods mostly, it should work anyway.

We also skip instrumentation when we're doing static verification.

Also we need to move some methods from geo_git_access to git_access,
because that's where it's trying to override methods, not in geo itself.

This also assumes GitAccessWiki does not use those methods, of course.
It won't work anyway, because it's not getting methods from GitAccess!