Allow shas that are nil to compare to nil shas
Previously, nil == nil was being treated as false. This is both clearly
wrong and more seriously broke the equality contract of `DiffRefs`,
since the same object would not compare equal to itself it any of its
components was nil.

Adds a test for commit.diff_refs equality, the motivating case.