Call openPendingTab only if open file has change
Now that staging and unstaging might result in absence of a change
when the changes cancel each other out, we need to add if condition
checks before opening a tab for a file to see if there are actually
any changes for that file or not. For example, adding a word "hello"
in staged and removing the same word in unstaged.