Add Operator dropdown
Add third token in the filtered_search -- the middle token which is
an operator with one of two permissible values. Equal or Not Equal
written as `=` or `!=`.

This operator allows an user to filter issues much more efficiently
wherein they can say `author=me assignee!=someone`, to exactly pin point
the search.