Adjust cache-repo CI stage run conditions
The cache-repo CI stage was failing on gitlab-org/gitlab-foss because
the CI_REPO_CACHE_CREDENTIALS wasn't defined.

We now loosen the restriction of this CI stage to run under two
conditions:

1. On a schedule
2. When CI_REPO_CACHE_CREDENTIALS is defined.

That way, dev.gitlab.org and other instances can produce their own
caches if necessary.

Also this changes the GIT_STRATEGY to none to reduce the overall size of
the pack files and avoid the pre-clone step. The current cache crept up
to close to a 1 GB as a result of an incremental update.