Add back Gitaly feature flag for cache invalidation
Part of https://gitlab.com/gitlab-org/gitaly/issues/2292

Related to https://gitlab.com/gitlab-org/gitaly/issues/2288

This change is essentially a revert of
https://gitlab.com/gitlab-org/gitlab/merge_requests/18766. Due to an
issue found in production (see
https://gitlab.com/gitlab-com/gl-infra/production/issues/1501#note_262058610),
the cache invalidator needs to be disabled until the cache walker is
able to remove empty directories. This feature flag will allow us to
reenable the cache invalidator when we are ready to test the updated
cache walker.