Gracefully handle CI lint errors in artifacts section
If an `artifacts` section is included in the CI lint as an array instead
of a Hash, users would see 500 errors when linting the file. Even though
we separate validation that checks the type, this validation doesn't
halt other validations from running. Since `expose_as_present?` is
called inside a validation, we need to perform a redundant check to
ensure the right type.

Closes https://gitlab.com/gitlab-org/gitlab/issues/194069