Do not sync stable branches for public commits
If a new commit is pushed to a stable branch on canonical, it will
trigger push mirroring to security.

Now we have a race where both canonical and security wants to trigger a
FOSS merge train ending up into a diverged branch.

This commit will check that the current commit is not available on
canonical before triggering a security FOSS merge train